+++ src/node.cpp
@@ -957,6 +957,7 @@ void CodeCompNode::dump(NodeDumper &dumper) const {
 
 #define EACH_ENUM(OP)                                                                              \
   OP(VAR)                                                                                          \
+  OP(VAR_IN_CMD_ARG)                                                                               \
   OP(MEMBER)                                                                                       \
   OP(TYPE)